api ファイル整理
