+++ app/src/main/java/com/example/funcy_portfolio_android/network/ConnectFuncyApi.kt
@@ -1,33 +1,22 @@
+package com.example.funcy_portfolio_android.network
 
-import com.example.funcy_portfolio_android.model.data.WorkData
 import com.google.gson.GsonBuilder
-
-
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Header
 
-//FuncyAPIサーバーの作品一覧取得URL
-private const val SERVER_URL = "http://192.168.50.252:8080/"
+val URL = "http://192.168.3.17:9000/"
 
-//gsonオブジェクトの作成
-private val gson = GsonConverterFactory.create(GsonBuilder().serializeNulls().create())
+val gson = GsonBuilder()
+    .serializeNulls()
+    .create()
 
-//Retrofitオブジェクトの作成
 private val retrofit = Retrofit.Builder()
-    .addConverterFactory(gson)
-    .baseUrl(SERVER_URL)
+    .baseUrl(URL)
+    .addConverterFactory(GsonConverterFactory.create(gson))
     .build()
 
-//FuncyAPIサーバーとの通信方法を定義
-interface ConnectFuncyApi{
-    @GET("works/10") // 取得する作品数
-    suspend fun getWorks(
-        @Header("token") token:String
-    ):List<WorkData>
-}
-
-object FuncyApi{
-    val retrofitService: ConnectFuncyApi by lazy {retrofit.create(ConnectFuncyApi::class.java)}
+object apiService {
+    val service: ICreationApi by lazy {
+        retrofit.create(ICreationApi::class.java)
+    }
 }
