+++ app/src/main/java/com/example/funcy_portfolio_android/ui/authentication/AuthenticationViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.funcy_portfolio_android.model.apiService
 import com.example.funcy_portfolio_android.model.data.AuthData
+import com.example.funcy_portfolio_android.network.apiService
 import kotlinx.coroutines.launch
 import retrofit2.HttpException
